refactor: update package declaration and imports for HttpHeaders move
