+++ application/src/ext-test/java/org/opentripplanner/ext/vehicleparking/bikeep/BikeepUpdaterTest.java
@@ -7,7 +7,7 @@ import java.time.Duration;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.test.support.ResourceLoader;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 
 class BikeepUpdaterTest {
