+++ application/src/main/java/org/opentripplanner/standalone/config/routerconfig/updaters/HttpHeadersConfig.java
@@ -2,7 +2,7 @@ package org.opentripplanner.standalone.config.routerconfig.updaters;
 
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 import org.opentripplanner.standalone.config.framework.json.OtpVersion;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 
 public class HttpHeadersConfig {
