+++ application/src/main/java/org/opentripplanner/updater/alert/gtfs/GtfsRealtimeAlertsUpdaterParameters.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.updater.alert.gtfs;
 
 import java.time.Duration;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.spi.PollingGraphUpdaterParameters;
 
 public record GtfsRealtimeAlertsUpdaterParameters(
