+++ application/src/main/java/org/opentripplanner/updater/trip/siri/updater/SiriETHttpTripUpdateSource.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 import javax.annotation.Nullable;
 import org.opentripplanner.framework.io.OtpHttpClientException;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.support.siri.SiriLoader;
 import org.opentripplanner.updater.trip.UpdateIncrementality;
 import org.opentripplanner.utils.tostring.ToStringBuilder;
