+++ application/src/main/java/org/opentripplanner/updater/trip/siri/updater/lite/SiriETLiteHttpTripUpdateSource.java
@@ -6,7 +6,7 @@ import java.net.URI;
 import java.time.Duration;
 import java.util.Optional;
 import org.opentripplanner.framework.io.OtpHttpClientException;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.support.siri.SiriLoader;
 import org.opentripplanner.updater.trip.UpdateIncrementality;
 import org.opentripplanner.updater.trip.siri.updater.EstimatedTimetableSource;
