+++ application/src/main/java/org/opentripplanner/updater/trip/siri/updater/lite/SiriETLiteUpdaterParameters.java
@@ -2,7 +2,7 @@ package org.opentripplanner.updater.trip.siri.updater.lite;
 
 import java.net.URI;
 import java.time.Duration;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.trip.siri.updater.SiriETUpdaterParameters;
 
 public record SiriETLiteUpdaterParameters(
