+++ application/src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/GbfsFeedLoaderImpl.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import javax.annotation.Nullable;
 import org.opentripplanner.framework.io.OtpHttpClient;
 import org.opentripplanner.framework.io.OtpHttpClientException;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.spi.UpdaterConstructionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
