+++ application/src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v2/GbfsFeedLoader.java
@@ -8,7 +8,7 @@ import org.mobilitydata.gbfs.v2_3.gbfs.GBFSFeed;
 import org.mobilitydata.gbfs.v2_3.gbfs.GBFSFeedName;
 import org.mobilitydata.gbfs.v2_3.gbfs.GBFSFeeds;
 import org.opentripplanner.framework.io.OtpHttpClient;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.spi.UpdaterConstructionException;
 import org.opentripplanner.updater.vehicle_rental.datasources.gbfs.GbfsFeedDetails;
 import org.opentripplanner.updater.vehicle_rental.datasources.gbfs.GbfsFeedLoaderImpl;
