+++ application/src/main/java/org/opentripplanner/updater/vehicle_rental/datasources/params/VehicleRentalDataSourceParameters.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.updater.vehicle_rental.datasources.params;
 
 import javax.annotation.Nullable;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.opentripplanner.updater.vehicle_rental.VehicleRentalSourceType;
 
 public interface VehicleRentalDataSourceParameters {
