+++ application/src/test/java/org/opentripplanner/updater/alert/siri/lite/SiriLiteHttpLoaderTest.java
@@ -6,7 +6,7 @@ import java.time.Duration;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.test.support.ResourceLoader;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import uk.org.siri.siri21.NaturalLanguageStringStructure;
 
 class SiriLiteHttpLoaderTest {
