+++ application/src/test/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v2/GbfsFeedLoaderTest.java
@@ -30,7 +30,7 @@ import org.mobilitydata.gbfs.v2_3.vehicle_types.GBFSVehicleType;
 import org.mobilitydata.gbfs.v2_3.vehicle_types.GBFSVehicleTypes;
 import org.opentripplanner.framework.io.OtpHttpClient;
 import org.opentripplanner.framework.io.OtpHttpClientFactory;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.slf4j.LoggerFactory;
 
 /**
