+++ application/src/test/java/org/opentripplanner/updater/vehicle_rental/datasources/gbfs/v3/GbfsFeedLoaderTest.java
@@ -15,7 +15,7 @@ import org.mobilitydata.gbfs.v3_0.vehicle_types.GBFSVehicleType;
 import org.mobilitydata.gbfs.v3_0.vehicle_types.GBFSVehicleTypes;
 import org.opentripplanner.framework.io.OtpHttpClient;
 import org.opentripplanner.framework.io.OtpHttpClientFactory;
-import org.opentripplanner.updater.spi.HttpHeaders;
+import org.opentripplanner.framework.io.HttpHeaders;
 import org.slf4j.LoggerFactory;
 
 /**
